requirements/requirments.txt
aiohttp==3.8.4
aiosignal==1.3.1
alpaca-py==0.13.4
alpha-vantage==2.3.1
annotated-types==0.6.0
antlr4-python3-runtime==4.9.3
anyio==3.7.1
appdirs==1.4.4
async-timeout==4.0.2
attrs==22.2.0
beautifulsoup4==4.12.1
bs4==0.0.1
bt==1.0.0
certifi==2022.12.7
cffi==1.15.1
cfgv==3.4.0
charset-normalizer==3.1.0
click==8.1.7
colorama==0.4.6
contourpy==1.0.7
cryptography==40.0.1
cssselect==1.2.0
cycler==0.11.0
decorator==5.1.1
distlib==0.3.8
exceptiongroup==1.2.0
fake-useragent==1.1.3
fastapi==0.105.0
feedparser==6.0.10
ffn==1.0.1
filelock==3.13.1
fonttools==4.39.3
frozendict==2.3.7
frozenlist==1.3.3
gevent==23.9.1
greenlet==3.0.2
h11==0.14.0
html5lib==1.1
hydra-core==1.3.2
identify==2.5.33
idna==3.4
importlib-metadata==6.1.0
importlib-resources==5.12.0
Jinja2==3.1.2
joblib==1.2.0
kiwisolver==1.4.4
lxml==4.9.2
MarkupSafe==2.1.2
matplotlib==3.7.1
mpmath==1.3.0
msgpack==1.0.7
multidict==6.0.4
multitasking==0.0.11
networkx==3.1
nodeenv==1.8.0
numpy==1.24.2
omegaconf==2.3.0
packaging==23.1
pandas==2.0.0
pandas-datareader==0.10.0
parse==1.19.0
peewee==3.17.0
Pillow==9.5.0
platformdirs==4.1.0
pre-commit==3.6.0
pycparser==2.21
pydantic==2.5.2
pydantic_core==2.14.5
pyee==8.2.2
pyparsing==3.0.9
pyppeteer==0.0.25
PyPrind==2.11.3
pyquery==2.0.0
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0.1
requests==2.31.0
requests-html==0.10.0
scikit-learn==1.2.2
scipy==1.10.1
sgmllib3k==1.0.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.4
sseclient-py==1.8.0
starlette==0.27.0
sympy==1.11.1
tabulate==0.9.0
threadpoolctl==3.1.0
tqdm==4.65.0
typing_extensions==4.9.0
tzdata==2023.3
urllib3==1.26.15
uvicorn==0.18.3
virtualenv==20.25.0
w3lib==2.1.1
webencodings==0.5.1
websocket-client==1.7.0
websockets==11.0.3
yarl==1.8.2
yfinance==0.2.35
zipp==3.15.0
zope.event==5.0
zope.interface==6.1
